+++ internal/controller/middleware/requestid/README.md
@@ -1,0 +1,42 @@
+# RequestID Middleware
+
+概要: このディレクトリは、リクエストIDを生成および管理するためのミドルウェアを提供します。リクエストIDは、各リクエストを一意に識別するために使用されます。
+
+## 役割
+
+- 各リクエストに一意のIDを付与し、リクエストのトラッキングを容易にします。
+- ログやデバッグ時にリクエストを特定するための手段を提供します。
+
+## 必要度
+
+### 本番運用での必須度
+
+- 必須度: 1
+  - 1: 本番運用で必須
+  - 2: 本番運用で推奨
+  - 3: 開発環境でのみ必要
+  - 4: テスト環境でのみ必要
+  - 5: 無し
+
+理由: 本番環境では、リクエストのトラッキングやデバッグのためにリクエストIDが必要不可欠です。
+
+### 開発/テスト運用での必須度
+
+- 必須度: 1
+  - 1: 開発/テスト運用で必須
+  - 2: 開発/テスト運用で推奨
+  - 3: 開発/テスト運用で任意
+  - 4: 無し
+
+理由: 開発やテスト環境でも、リクエストのトラッキングやデバッグを効率的に行うために必要です。
+
+### 無効化した場合の影響
+
+- リクエストのトラッキングが困難になり、ログやデバッグ時に特定のリクエストを識別することができなくなります。
+- 複数のリクエストが混在する環境での問題解決が難しくなります。
+
+## 注意点
+
+- リクエストIDは一意である必要があります。
+- セキュリティ上の理由から、リクエストIDに機密情報を含めないようにしてください。
+- 他のミドルウェアと組み合わせて使用する場合、リクエストIDが適切に設定されていることを確認してください。
