+++ internal/controller/middleware/requestid/request_id.go
@@ -1,0 +1,19 @@
+// Package requestid は、リクエストIDミドルウェアをラップするためのパッケージです。
+package requestid
+
+import (
+	"github.com/labstack/echo/v4"
+	"github.com/labstack/echo/v4/middleware"
+)
+
+// Middleware は、リクエストIDを生成するミドルウェアを返します。
+func Middleware() echo.MiddlewareFunc {
+	return middleware.RequestID()
+}
+
+// GetRequestIDFromResponse は、レスポンスのEchoコンテキストからリクエストIDを取得します。
+func GetRequestIDFromResponse(
+	c echo.Context,
+) string {
+	return c.Response().Header().Get(echo.HeaderXRequestID)
+}
